Refactor rename Collection module to Payments module
