+++ src/ADempiere/modules/pos/index.ts
@@ -3,7 +3,7 @@ export * from './POSType'
 export * from './POSService'
 // VUEX
 export * from './store/PointOfSales'
-export * from './store/Collection'
+export * from './store/Payments'
 export * from './store/KeyLayout'
 export * from './store/Order'
 export * from './store/OrderLines'
